Move low level files to Core folder.
