+++ GridGames/ScratchCard/Hammerpot/hammerpot.py
@@ -1,8 +1,8 @@
+from Core.helper_functions import *
 from GridGames.ScratchCard.Hammerpot.feedback import HammerpotFeedback
 from GridGames.ScratchCard.Hammerpot.render_hammerpot import RenderHammerpot
 from GridGames.ScratchCard.constants import *
 from GridGames.ScratchCard.scratch_card import ScratchCard
-from helper_functions import *
 
 
 class Hammerpot(ScratchCard):
