+++ GridGames/ScratchCard/scratch_card.py
@@ -1,7 +1,7 @@
+from helper_functions import *
+
 from GridGames.ScratchCard.constants import *
 from GridGames.grid_game_class import GridGame
-from helper_functions import *
-from joinable_game_class import JoinableGame
 
 
 class ScratchCard(GridGame):
