+++ Managers/GameManagers/game_manager.py
@@ -1,5 +1,4 @@
 import asyncio
-from constants import *
 
 
 class GameManager:
